renderer: extract shared sink and renderer protocol
The upcoming TTY renderer needs the same per-build log sink
(BuildOutput) and lifecycle interface (start/finish/abort) as the CI
renderer. Move them plus fmt_duration into renderer.py; BuildOutput
talks to its owner through a Host protocol (clock + live line
emission) instead of depending on CIRenderer directly.
